[PRODUCT] feat: 재고 수정 API & map struct 라이브러리 적용

+++ micro-commerce-product/src/main/java/com/microcommerce/product/domain/dto/req/CreateProductReqDto.java
@@ -3,9 +3,9 @@ package com.microcommerce.product.domain.dto.req;
 public record CreateProductReqDto(
         Long sellerId,
         String name,
-        Long price,
+        Integer price,
         String category,
         String description,
-        Long stock
+        Integer stock
 ) {
 }
